moved methods to ReaderNodeUtil class
Also moved tests to ReaderNodeUtilTest class
